// src/components/checkbox/ub-checkbox.d.ts
import { LitElement } from "lit";
export declare class UbCheckbox extends LitElement {
  private _checked;
  value: string;
  name: string | undefined;
  set checked(val: boolean);
  get checked(): boolean;
  indeterminate: boolean;
  disabled: boolean;
  input: HTMLInputElement;
  static styles: CSSStyleSheet[];
  protected internals: ElementInternals;
  static formAssociated: boolean;
  constructor();
  private handleOnChange;
  render(): import("lit-html").TemplateResult<1>;
}
declare global {
  interface HTMLElementTagNameMap {
    "ub-checkbox": UbCheckbox;
  }
}
//# sourceMappingURL=ub-checkbox.d.ts.map